+++ libnavigation-base/src/test/java/com/mapbox/navigation/base/route/RouteExclusionsTest.kt
@@ -25,7 +25,7 @@ class RouteExclusionsTest {
     fun setup() {
         mockkObject(NativeRouteParserWrapper)
         every {
-            NativeRouteParserWrapper.parseDirectionsResponse(any(), any())
+            NativeRouteParserWrapper.parseDirectionsResponse(any(), any(), any())
         } answers {
             val routesCount = JSONObject(this.firstArg<String>())
                 .getJSONArray("routes")